Add confirmation page

+++ src/App.js
@@ -1,7 +1,8 @@
 import { useRef, useState, useEffect } from "react";
-import Card from "./Components/Card/Card"
-import Form from "./Components/Form/Form"
-import Nav from "./Components/Nav/Nav"
+import Card from "./Components/Card/Card";
+import Form from "./Components/Form/Form";
+import Nav from "./Components/Nav/Nav";
+import ConfirmationMod from "./Components/ConfirmationMod/ConfirmationMod";
 import apiService from "./lib/api-service";
 import './App.css';
 
@@ -14,6 +15,8 @@ function App() {
   const [token, setToken] = useState("");
   const [edit, setEdit] = useState(false);
   const [orderSet, setOrderSet] = useState("");
+  const [showConfirmMod, setShowConfirmMod] = useState(false);
+  const [toDelete, setToDelete] = useState("");
   const inputRef = useRef(null)
 
   useEffect(() => {
@@ -54,15 +57,28 @@ function App() {
       .catch(err => console.log(err))
 
   }
+  let interval = 0;
+  const askConfDeleteCart = (idCard) => {
+    setToDelete(idCard)
+    setShowConfirmMod(true);
+  }
 
   const deleteCard = (idCard) => {
+    if (!idCard) {
+      setShowConfirmMod(false);
+      return;
+    }
     apiService.deleteCard(idCard, token)
       .then(res => {
         const newCards = [...cards.filter(el => el.id !== idCard)];
         setCards([...newCards]);
         setCardsOrig([...newCards]);
+        setShowConfirmMod(false);
       })
-      .catch(err => console.log(err))
+      .catch(err => {
+        console.log(err);
+        setShowConfirmMod(false);
+      });
   }
 
   const updateCard = (card_id, formData) => {
@@ -123,6 +139,12 @@ function App() {
         break;
     }
   }
+
+  const globalClick = (e) => {
+    setEdit && setEdit(false);
+    setCreate && setCreate(false);
+  }
+
   return (
     <div className="App">
       <Nav orderArr={orderArr} orderSet={orderSet} />
@@ -133,12 +155,12 @@ function App() {
         </div>
         {
           cards && cards.map(card =>
-            <Card key={card.id} name="card" openEditForm={openEditForm} card={card} deleteCard={deleteCard} updateCard={updateCard} openEditForm={openEditForm} />
+            <Card key={card.id} name="card" openEditForm={openEditForm} card={card} askConfDeleteCart={askConfDeleteCart} updateCard={updateCard} openEditForm={openEditForm} />
           )
         }
       </div>
       {create ?
-        <Form setCreate={setCreate} addCard={addCard} deleteCard={deleteCard} />
+        <Form setCreate={setCreate} globalClick={globalClick} addCard={addCard} deleteCard={deleteCard} />
         :
         <div className="addDiv">
           <button data-testid="buttonAdd" className="buttonAdd" onClick={() => setCreate(true)}></button>
@@ -146,6 +168,9 @@ function App() {
       }
       {
         edit && <Form card={cardToEdit} edit setEdit={setEdit} updateCard={updateCard} />
+      }
+      {
+        showConfirmMod && <ConfirmationMod deleteCard={deleteCard} idCard={toDelete} globalClick={globalClick} />
       }
     </div>
   );
